chore: Merge start pages (unlock and get started)
