--- source/pages/Start-auth/index.tsx
@@ -1,1 +1,0 @@
-export * from './Start';
